Enable stubbing of DTUbasen and CAS in development mode
To enable toshokan to run without a network connection (or without
access to the DTU network) it is now  possible to stub out DTUbasen
and CAS by setting

    Toshokan::Application.config.stub_authentication=true

(e.g. in config/application.local.rb). This makes OmniAuth and
Dtubase cooperate to always log in as a user with identifier "0"